Move a bunch of files
