+++ src/commands/index.ts
@@ -1,0 +1,5 @@
+import { runDiff } from './diff'
+
+export async function runCommand() {
+  await runDiff()
+}
