+++ src/utils/message.ts
@@ -1,5 +1,5 @@
 import { getInput } from '@actions/core'
-import { ResourceDiff } from './diff'
+import { ResourceDiff } from './types'
 
 export const credits =
   '_Generated by [Locize Diff](https://github.com/Widen/locize-diff)._'
