+++ src/utils/types.ts
@@ -4,3 +4,8 @@ export interface ResourceCollection {
   key: string
   resources: Dict
 }
+
+export interface ResourceDiff {
+  key: string
+  diffs: Record<string, { left?: string; right?: string }>
+}
